+++ src/main/java/com/coniverse/dangjang/domain/guide/bloodsugar/service/BloodSugarGuideGenerateService.java
@@ -8,8 +8,8 @@ import com.coniverse.dangjang.domain.code.enums.CommonCode;
 import com.coniverse.dangjang.domain.code.enums.GroupCode;
 import com.coniverse.dangjang.domain.guide.bloodsugar.document.BloodSugarGuide;
 import com.coniverse.dangjang.domain.guide.bloodsugar.document.SubGuide;
-import com.coniverse.dangjang.domain.guide.bloodsugar.guideformat.GuideFormatFactory;
-import com.coniverse.dangjang.domain.guide.bloodsugar.guideformat.format.GuideFormat;
+import com.coniverse.dangjang.domain.guide.bloodsugar.factory.GuideFormatFactory;
+import com.coniverse.dangjang.domain.guide.bloodsugar.factory.guideformat.GuideFormat;
 import com.coniverse.dangjang.domain.guide.bloodsugar.mapper.BloodSugarGuideMapper;
 import com.coniverse.dangjang.domain.guide.bloodsugar.repository.BloodSugarGuideRepository;
 import com.coniverse.dangjang.domain.guide.common.dto.GuideResponse;
